Fixed bad assertions, tests will now run using maven
