+++ tests-arquillian/src/test/java/org/jboss/weld/tests/specialization/modular/InjectedBean1.java
@@ -22,12 +22,12 @@ import javax.inject.Inject;
 public class InjectedBean1 {
 
     @Inject
-    private Factory factory;
+    private Instance<Factory> factory;
 
     @Inject
     private Instance<Product> product;
 
-    public Factory getFactory() {
+    public Instance<Factory> getFactory() {
         return factory;
     }
